+++ apps/native-component-list/src/screens/UI/UIScreen.tsx
@@ -84,6 +84,14 @@ export const UIScreens = [
       return optionalRequire(() => require('./ProgressScreen'));
     },
   },
+  {
+    name: 'Stepper component',
+    route: 'ui/stepper',
+    options: {},
+    getComponent() {
+      return optionalRequire(() => require('./StepperScreen'));
+    },
+  },
 ];
 if (Platform.OS === 'ios') {
   UIScreens.push({